requirements.txt
opencv-python
ultralytics
supervision
numpy
argparse
torch
python-dotenv
websocket-client
rel